Refactoring and made a clear seperation between front and backend
