+++ package.json
@@ -22,7 +22,6 @@
     "@types/fs-extra": "^5.0.4",
     "@types/react": "^16.7.18",
     "@types/react-dom": "^16.0.11",
-    "@types/react-sidebar": "^3.0.0",
     "css-loader": "^2.1.0",
     "electron": "^4.0.0",
     "file-loader": "^3.0.1",
@@ -40,8 +39,6 @@
     "idb": "^3.0.2",
     "mobx-react": "^5.4.3",
     "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-grid-gallery": "^0.5.3",
-    "react-sidebar": "^3.0.2"
+    "react-dom": "^16.7.0"
   }
 }
