+++ src/renderer/frontend/components/Sidebar.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import TagList from './TagList';
+
+export default ({ children }: { children: JSX.Element[] }) => (
+  <div>
+   {children}
+  </div>
+);
