+++ tslint.json
@@ -14,7 +14,8 @@
     "object-literal-sort-keys": false,
     "max-classes-per-file": false,
     "no-empty-interface": false,
-    "indent": [true, "spaces", 2]
+    "indent": [true, "spaces", 2],
+    "member-access": [false, "warning"]
   },
   "jsRules": {
     "max-line-length": {
